requirement.txt
djangorestframework       3.13.1
Django                    2.2